feat: add recoil comp
